refactor(geotiff): move source middleware into src/source/
Group SourceReadaheadCache and its mutex() helper under a source/
subdirectory to make the chunkd Source middleware layer explicit. No
behavior change.
