+++ packages/geotiff/src/source/concurrency.ts
