+++ packages/geotiff/src/source/readahead-cache.ts
